Made it compile of 64 bit machines
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@645 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ apps/samsonLogServer/main_samsonLogServer.cpp
@@ -256,12 +256,12 @@ int main(int argC, const char *argV[])
 	else
 		LM_T(LmtInit, ("Not connecting to Controller"));
 
-	LM_T(LmtInit, ("Number of spawners: %d", (int) spawnerList[0]));
-	if ((int) spawnerList[0] != 0)
+	LM_T(LmtInit, ("Number of spawners: %d", (long int) spawnerList[0]));
+	if ((long int) spawnerList[0] != 0)
 	{
 		int ix;
 
-		for (ix = 1; ix <= (int) spawnerList[0]; ix++)
+		for (ix = 1; ix <= (long int) spawnerList[0]; ix++)
 		{
 			int fd;
